+++ src/euphorie/client/config.py
@@ -1,3 +1,6 @@
 FULL_ACCOUNT = "full"
 GUEST_ACCOUNT = "guest"
 CONVERTED_ACCOUNT = "converted"
+
+LOCKING_SET_ACTIONS = ("lock_set", "lock_refresh")
+LOCKING_ACTIONS = ("lock_unset",) + LOCKING_SET_ACTIONS